integrate with ielts sample doTest page

+++ testservice/src/main/java/com/lingo/testservice/model/Test.java
@@ -25,7 +25,7 @@ public class Test implements Serializable {
     long id;
     @Column(unique = true)
     String title;
-    int maxScore;
+    double maxScore;
     int timeLimit;
     @Enumerated(EnumType.STRING)
     TestType type;
